Refactor code to remove unnecessary context.Database.EnsureCreated() calls and update assertions to use Shouldly for consistency in tests.
