+++ Examples/OrderService/OrderService.Boa/OrderServiceBoaTestFactoryFixture.cs
@@ -3,11 +3,8 @@ using DataPreparation.Factory.Testing;
 using DataPreparation.Provider;
 using DataPreparation.Testing;
 using DataPreparation.Testing.Factory;
-using FluentAssertions;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Moq;
+using Microsoft.Extensions.Logging;
 using OrderService.Boa.OrderService.Abilities;
 using OrderService.Boa.OrderService.Questions;
 using OrderService.Boa.OrderService.Tasks;
@@ -15,21 +12,20 @@ using OrderService.Boa.ShowCases.Factories;
 using OrderService.DTO;
 using OrderService.Models;
 using OrderService.Repository;
-using ILogger = NUnit.Framework.Internal.ILogger;
-using Microsoft.Extensions.Logging;
-using Range = Moq.Range;
+using Shouldly;
 
 namespace OrderService.Boa;
 
 [DataPreparationFixture]
-//TODO [Parallelizable(ParallelScope.All)]
-public class OrderServiceBoaTestFactoryFixture:IDataPreparationLogger, IDataPreparationTestServices
+[Parallelizable(ParallelScope.All)]
+public class OrderServiceBoaTestFactoryFixture : IDataPreparationLogger, IDataPreparationTestServices
 {
-    public  void DataPreparationServices(IServiceCollection serviceCollection)
+    public void DataPreparationServices(IServiceCollection serviceCollection)
     {
       
     }
-    public  ILoggerFactory InitializeDataPreparationTestLogger()
+    
+    public ILoggerFactory InitializeDataPreparationTestLogger()
     {
         return LoggerFactory.Create(builder =>
         {
@@ -37,7 +33,6 @@ public class OrderServiceBoaTestFactoryFixture:IDataPreparationLogger, IDataPrep
             builder.AddDebug();
             builder.AddConsole();
         });
-        
     }
     
     [DataPreparationTest]
@@ -58,22 +53,18 @@ public class OrderServiceBoaTestFactoryFixture:IDataPreparationLogger, IDataPrep
         // Act
         var createTask = CreateOrderTask.For(orderDto); //task that will be executed CreateOrder  -  get customer by ID and insert order
         actor.AttemptsTo(createTask);
-        Order result = actor.AsksFor( OrderById.WithId(createTask.CreatedOrder.Id)); // there can be find mock for GetByID, or add real test data to DB
+        Order result = actor.AsksFor(OrderById.WithId(createTask.CreatedOrder.Id)); // there can be find mock for GetByID, or add real test data to DB
 
-        
-        // Assert
-        result.Should().NotBeNull();
-        result.CustomerId.Should().Be(orderDto.CustomerId);
-        result.OrderItems.Should().NotBeNull();
-        result.OrderItems.Should().HaveCount( (int)PreparationContext.GetFactory().Was<OrderItem,OrderItemFactory>().ToList().Count);
-      
+        // Assert - converted to Shouldly
+        result.ShouldNotBeNull();
+        result.CustomerId.ShouldBe(orderDto.CustomerId);
+        result.OrderItems.ShouldNotBeNull();
+        result.OrderItems.Count().ShouldBe((int)PreparationContext.GetFactory().Was<OrderItem,OrderItemFactory>().ToList().Count);
     }
-    
     
     [DataPreparationTest]
     public async Task CreateOrder_FullOrderDTO_ReturnsOrderFactoryAsync() // Show Case for Test Data Preparation
     {
-      
         // Arrange
         var customer = await PreparationContext.GetFactory().NewAsync<Customer,CustomerFactoryAsync>();
         OrderDTO orderDto = await PreparationContext.GetFactory().GetAsync<OrderDTO,OrderDtoFactoryAsync>();
@@ -89,15 +80,12 @@ public class OrderServiceBoaTestFactoryFixture:IDataPreparationLogger, IDataPrep
         // Act
         var createTask = CreateOrderTask.For(orderDto); //task that will be executed CreateOrder  -  get customer by ID and insert order
         actor.AttemptsTo(createTask);
-        Order result = actor.AsksFor( OrderById.WithId(createTask.CreatedOrder.Id)); // there can be find mock for GetByID, or add real test data to DB
+        Order result = actor.AsksFor(OrderById.WithId(createTask.CreatedOrder.Id)); // there can be find mock for GetByID, or add real test data to DB
 
-        
-        // Assert
-        result.Should().NotBeNull();
-        result.CustomerId.Should().Be(orderDto.CustomerId);
-        result.OrderItems.Should().NotBeNull();
-        result.OrderItems.Should().HaveCount( PreparationContext.GetFactory().Was<OrderItem,OrderItemFactoryAsync>().ToList().Count);
-      
+        // Assert - converted to Shouldly
+        result.ShouldNotBeNull();
+        result.CustomerId.ShouldBe(orderDto.CustomerId);
+        result.OrderItems.ShouldNotBeNull();
+        result.OrderItems.Count().ShouldBe(PreparationContext.GetFactory().Was<OrderItem,OrderItemFactoryAsync>().ToList().Count);
     }
-    
 }
